+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 CHANGELOG
 =========
 
-## 0.1.1 (Unreleased)
+## 0.2.1 (Unreleased)
+
+## 0.2.0 (2017-03-20, beta release)
+
+#### features:
+ - added LazyCat.Video()
+
+#### fixes:
+ - fixed adding background images multiple times on one element.
 
 ## 0.1.0 (2017-03-02, beta release)